Implemented OnboardingNavigator.tsx
