+++ native/App.tsx
@@ -1,34 +1,9 @@
-import { NavigationContainer } from "@react-navigation/native";
-import { createStackNavigator } from "@react-navigation/stack";
 import React from "react";
 import { StyleSheet } from "react-native";
-import { OnboardingCompleted } from "./src/feature/screens/OnboardingCompleted";
-import { OnboardingPhoneVerification } from "./src/feature/screens/OnboardingPhoneVerification";
-import { OnboardingStoreConnection } from "./src/feature/screens/OnboardingStoreConnection";
-import { OnboardingWelcome } from "./src/feature/screens/OnboardingWelcome";
-
-const Stack = createStackNavigator();
+import { OnboardingNavigator } from "./src/navigation/navigator/OnboardingNavigator";
 
 const App: React.FC<{}> = () => {
-	return (
-		<NavigationContainer>
-			<Stack.Navigator>
-				<Stack.Screen name="OnboardingWelcome" component={OnboardingWelcome} />
-				<Stack.Screen
-					name="OnboardingPhoneVerification"
-					component={OnboardingPhoneVerification}
-				/>
-				<Stack.Screen
-					name="OnboardingStoreConnection"
-					component={OnboardingStoreConnection}
-				/>
-				<Stack.Screen
-					name="OnboardingCompleted"
-					component={OnboardingCompleted}
-				/>
-			</Stack.Navigator>
-		</NavigationContainer>
-	);
+	return <OnboardingNavigator />;
 };
 
 export default App;
